# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "gutv_booker",
    "version": "1.0"
  },
  "paths": {
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Equipment/create_type": {
      "post": {
        "tags": [
          "Equipment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Equipment/get_type_by_id/{id}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              }
            }
          }
        }
      }
    },
    "/api/Equipment/get_type_by_name/{namePart}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "namePart",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Equipment/get_type_by_category/{category}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "category",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EquipmentCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EquipmentType"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Equipment/update_type/{id}": {
      "put": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Equipment/delete_type/{id}": {
      "delete": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Equipment/create_item": {
      "post": {
        "tags": [
          "Equipment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EquipmentItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentItem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Equipment/get_item_by_id/{id}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              }
            }
          }
        }
      }
    },
    "/api/Equipment/get_item_by_type/{typeId}": {
      "get": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "typeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EquipmentType"
                }
              }
            }
          }
        }
      }
    },
    "/api/Equipment/delete_item/{id}": {
      "delete": {
        "tags": [
          "Equipment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EquipmentCategory": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "format": "int32"
      },
      "EquipmentItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "equipmentTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "inventoryNumber": {
            "type": "string",
            "nullable": true
          },
          "available": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "EquipmentType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/EquipmentCategory"
          },
          "attributes": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "login": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}